lib/storage: move testing/synctest usage into separate files ending with _synctest_test.go
The files with synctest usage are built only if the goexperiment.synctest build tag is set.
This allows running `go test ./lib/...` and `go vet ./lib/...` without the need to pass GOEXPERIMENT=synctest to them.

This is a follow-up for d33d7e20bedf2674024bc43e9c0d9f799b941ee2
Updates https://github.com/VictoriaMetrics/VictoriaMetrics/pull/8848
